+++ UniversalNMS/src/main/java/com/acrylic/universalnms/pathfinder/jps/JPSPathNode.java
@@ -8,7 +8,7 @@ public class JPSPathNode extends JPSBaseNode {
 
     private final int facingX, facingZ;
 
-    protected JPSPathNode(@NotNull JPSBaseNode start, int x, int y, int z, int facingX, int facingZ) {
+    protected JPSPathNode(@NotNull JPSBaseNode start, float x, float y, float z, int facingX, int facingZ) {
         super(start.getPathfinder(), start, x, y, z);
         this.facingX = facingX;
         this.facingZ = facingZ;
